refactor: move get logic to provider
